+++ src/Service/StaticContent.php
@@ -25,15 +25,10 @@ class StaticContent
     /** @var QueryBuilderFactoryInterface */
     private $queryBuilderFactory;
 
-    /** @var ModuleSettings */
-    private $moduleSettings;
-
     public function __construct(
-        QueryBuilderFactoryInterface $queryBuilderFactory,
-        ModuleSettings $moduleSettings
+        QueryBuilderFactoryInterface $queryBuilderFactory
     ) {
         $this->queryBuilderFactory = $queryBuilderFactory;
-        $this->moduleSettings = $moduleSettings;
     }
 
     public function ensurePayPalPaymentMethods(): void
@@ -46,7 +41,6 @@ class StaticContent
             }
             $paymentMethod = oxNew(EshopModelPayment::class);
             if ($paymentMethod->load($paymentId)) {
-                $this->reActivatePaymentMethod($paymentId);
                 continue;
             }
             $this->createPaymentMethod($paymentId, $paymentDefinitions);
@@ -109,25 +103,9 @@ class StaticContent
         }
     }
 
-    protected function reActivatePaymentMethod(string $paymentId): void
-    {
-        $activePayments = $this->moduleSettings->getActivePayments();
-        if (!in_array($paymentId, $activePayments, true)) {
-            return;
-        }
-
-        /** @var EshopModelPayment $paymentModel */
-        $paymentModel = oxNew(EshopModelPayment::class);
-        $paymentModel->load($paymentId);
-
-        $paymentModel->oxpayments__oxactive = new Field(true);
-
-        $paymentModel->save();
-    }
-
     /**
      * Try to load payment model based on given id an set payment inactive
-     * 
+     *
      * @param string $paymentId
      * @return void
      * @throws \Exception